metanodes and tests
